リファクタリング
・おすすめ画面のMVPを整形
・Viewには表示処理のみ（show...やclose...のみ）
・ビジネスロジックをPresenterに集約
・remoteRepositoryの削除
・Repositoryはひとつのクラスに集約
