+++ app/src/main/java/com/kk/gourmetapp/recommend/RecommendPresenter.kt
@@ -12,27 +12,28 @@ import com.kk.gourmetapp.data.source.DataRepository
 import com.kk.gourmetapp.data.source.DataSource
 
 class RecommendPresenter(recommendView: RecommendContract.View, context: Context)
-    : RecommendContract.UserActionListener {
+    : RecommendContract.Presenter {
 
     // フラグメントはコンストラクタで受け取る
     private var mRecommendView: RecommendContract.View = recommendView
     // レポジトリはコンストラクタで受け取る
-    private var mDataRepository: DataRepository? = DataRepository(context)
+    private var mDataRepository: DataRepository = DataRepository(context)
 
     init {
         // FragmentにPresenterの登録要求
-        mRecommendView.setUserActionListener(this)
+        mRecommendView.setPresenter(this)
     }
 
     /**
-     * {@inheritDoc}
+     * ぐるなびのお店情報生成
+     * @param bundle 現在地
      */
     override fun createGurunaviInfo(bundle: Bundle) {
 
-        val isCelebMode: Boolean? = mDataRepository?.isCelebMode()
+        val isCelebMode: Boolean? = mDataRepository.isCelebMode()
 
             // Repository側でモデルクラス作成する
-            mDataRepository?.createGurunaviInfo(object : DataSource.CreateGurunaviShopCallback {
+            mDataRepository.createGurunaviInfo(object : DataSource.CreateGurunaviShopCallback {
                 override fun createGurunaviShop(
                     shops: MutableList<GurunaviShop>, imageLoader: ImageLoader?) {
                     // ぐるなびのレストラン情報が取得できた場合にはUI更新をかける
@@ -41,7 +42,7 @@ class RecommendPresenter(recommendView: RecommendContract.View, context: Context
 
                 override fun onError() {
                     // 結果が取得できなかったら通信状況確認ダイアログ表示
-                    if (!isConnectNetwork()) {
+                    if (!mDataRepository.isConnectNetwork()) {
                         mRecommendView.showNetworkErrorDialog()
                     }
                 }
@@ -49,24 +50,25 @@ class RecommendPresenter(recommendView: RecommendContract.View, context: Context
     }
 
     /**
-     * {@inheritDoc}
+     * ホットペッパーのお店情報生成
+     * @param bundle 現在地
      */
     override fun createHotpepperInfo(bundle: Bundle) {
 
-        val isCelebMode: Boolean? = mDataRepository?.isCelebMode()
+        val isCelebMode: Boolean? = mDataRepository.isCelebMode()
 
-        mDataRepository?.createHotpepperInfo(object :DataSource.CreateHotpepperShopCallback {
+        mDataRepository.createHotpepperInfo(object :DataSource.CreateHotpepperShopCallback {
             override fun createHotpepperShop(
                 shops: MutableList<HotpepperShop>, imageLoader: ImageLoader?) {
                 // ホットペッパーのレストラン情報が取得できた場合にはUI更新をかける
                 mRecommendView.showHotpepperShops(shops, imageLoader)
                 // 画像解析で取得保存しておいたキーワードを空文字にしておく
-                mDataRepository?.removeRecognizeKeyword()
+                mDataRepository.removeRecognizeKeyword()
             }
 
             override fun onError() {
                 // 結果が取得できなかったら通信状況確認ダイアログ表示
-                if (!isConnectNetwork()) {
+                if (!mDataRepository.isConnectNetwork()) {
                     mRecommendView.showNetworkErrorDialog()
                 }
             }
@@ -74,68 +76,71 @@ class RecommendPresenter(recommendView: RecommendContract.View, context: Context
     }
 
     /**
-     * {@inheritDoc}
+     * ぐるなびのクレジット画像の読み込み
+     * @return Web画像のRequest
      */
     override fun loadGurunaviCredit(): RequestBuilder<Drawable>? {
-        return mDataRepository?.loadGurunaviCredit()
+        return mDataRepository.loadGurunaviCredit()
     }
 
     /**
-     * {@inheritDoc}
+     * ホットペッパーのクレジット画像の読み込み
+     * @return Web画像のRequest
      */
     override fun loadHotpepperCredit(): RequestBuilder<Drawable>? {
-        return mDataRepository?.loadHotpepperCredit()
+        return mDataRepository.loadHotpepperCredit()
     }
 
     /**
-     * {@inheritDoc}
+     * ぐるなびのクレジット画像の遷移先読み込み
      */
-    override fun loadGurunaviCreditUri(): Uri? {
-        return mDataRepository?.loadGurunaviCreditUri()
+    override fun updateGurunaviCreditTransition() {
+        val uri: Uri =  mDataRepository.loadGurunaviCreditUri()
+        mRecommendView.showApiCreditTransition(uri)
     }
 
     /**
-     * {@inheritDoc}
+     * ホットペッパーのクレジット画像の遷移先読み込み
      */
-    override fun loadHotpepperCreditUri(): Uri? {
-        return mDataRepository?.loadHotpepperCreditUri()
+    override fun updateHotpepperCreditTransition(){
+        val uri: Uri = mDataRepository.loadHotpepperCreditUri()
+        mRecommendView.showApiCreditTransition(uri)
     }
 
     /**
-     * {@inheritDoc}
+     * セレブモードの更新
      */
-    override fun isCelebMode(): Boolean {
-        return mDataRepository?.isCelebMode() as Boolean
+    override fun updateCelebMode() {
+        if (mDataRepository.isCelebMode()) {
+            mRecommendView.showCelebBackground()
+        } else {
+            mRecommendView.removeCelebBackground()
+        }
     }
 
     /**
-     * {@inheritDoc}
+     * おすすめショップの更新可否
+     * @return true  : 更新必要
+     *         false : 更新不必要
      */
     override fun shouldUpdate(): Boolean  {
-        return mDataRepository?.shouldUpdate() as Boolean
+        return mDataRepository.shouldUpdate()
     }
 
     /**
-     * {@inheritDoc}
-     */
-    override fun isConnectNetwork(): Boolean {
-        return mDataRepository?.isConnectNetwork() as Boolean
-    }
-
-    /**
-     * {@inheritDoc}
+     * ショップ情報の読み込み処理
      */
     override fun loadShopInfo() {
-        if (mDataRepository!!.hasLocationPermission()) {
+        if (mDataRepository.hasLocationPermission()) {
             // パーミッションが許可されている場合には現在地を取得する
-            mDataRepository?.getCurrentLocation(object : DataSource.LocationCallback {
+            mDataRepository.getCurrentLocation(object : DataSource.LocationCallback {
                 override fun onComplete(bundle: Bundle) {
                     // 現在地取得後にショップ情報を作成
                     createGurunaviInfo(bundle)
                     createHotpepperInfo(bundle)
 
                     // 現在地取得中画面を閉じる
-                    mRecommendView.removeLoadingFragment()
+                    mRecommendView.closeLocationLoadingFragment()
                 }
             })
         } else {
@@ -145,34 +150,29 @@ class RecommendPresenter(recommendView: RecommendContract.View, context: Context
     }
 
     /**
-     * {@inheritDoc}
+     * すでに取得済みの現在地取得
+     * @return 現在地
      */
     override fun getSavedCurrentLocation(): Bundle {
-        return mDataRepository!!.getSavedCurrentLocation()
+        return mDataRepository.getSavedCurrentLocation()
     }
 
     /**
-     * {@inheritDoc}
-     */
-    override fun saveManualKeyword(keyword: String) {
-        mDataRepository?.saveManualKeyword(keyword)
-    }
-
-    /**
-     * {@inheritDoc}
+     * 手動入力キーワードによる再検索処理
+     * @param keyword 手動入力キーワード
      */
     override fun researchShopManualKeyword(keyword: String) {
         // キーワードをPreferenceに一時保存
-        mDataRepository?.saveManualKeyword(keyword)
+        mDataRepository.saveManualKeyword(keyword)
 
         // 画像解析が完了している場合には取得済みの現在地で再検索をかける
-        val bundle: Bundle? = mDataRepository!!.getSavedCurrentLocation()
+        val bundle: Bundle? = mDataRepository.getSavedCurrentLocation()
         if (bundle != null) {
             createGurunaviInfo(bundle)
             createHotpepperInfo(bundle)
         }
 
         // Preferenceに一時保存していたキーワードを削除する
-        mDataRepository?.clearManualKeyword()
+        mDataRepository.clearManualKeyword()
     }
 }
